Register "Select->Spectral"; Break SelectMenus -> SpectrumAnalyst
